# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "My API Documentation",
    "description": "My API description goes here.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost"
    }
  ],
...
  "paths": {
...
    "/api/v1/products/{productId}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get product by ID",
        "description": "Get product information by their ID.",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of the product to retrieve.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1nt3io20_3"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "error": false,
                  "status": 200,
                  "message": "This product is available",
                  "body": [
                    {
                      "category_id": 1,
                      "name": "Smartphone X",
                      "color": "Black",
                      "description": "A high-quality smartphone.",
                      "price": 50,
                      "quantity": 10,
                      "image": "smartphone_x.jpg"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "User not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Deactivates a product so it can't be bought, by it's product ID (note that the product itself won't be deleted).",
        "description": "Deactivates a product so it can't be bought, by it's product ID (note that the product itself won't be deleted).",
        "operationId": "deactivateProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of the product to deactivate.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseWithStringBody"
                },
                "example": {
                  "error": false,
                  "status": 200,
                  "message": "Product deactivated",
                  "body": "The item you wanted to update was indeed updated."
                }
              }
            }
          },
          "404": {
            "description": "Product was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          },
          "500": {
            "description": "Various errors might ocurr.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                },
                "examples": {
                  "InternalError_1": {
                    "summary": "The product you tried to deactivate was already inactive and no update was made.",
                    "value": {
                      "error": true,
                      "status": 500,
                      "message": "No update was made to the product because it has the same state."
                    }
                  },
                  "BadRequest_2": {
                    "summary": "Internal server error.",
                    "value": {
                      "error": true,
                      "status": 500,
                      "message": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "description": "Indicates whether an error occurred or not.",
            "example": false
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code.",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "A message describing the response.",
            "example": "This request was successfull, here is some relevant information about this request:"
          },
          "body": {
            "type": "array",
            "description": "An array of objects that can vary depending on the specific request. For example, when doing request to getUsers you will receive and object with their email, username and avatar.",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "SuccessResponseWithStringBody": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "description": "Indicates whether an error occurred or not.",
            "example": false
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code.",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "A message describing the response.",
            "example": "This request was successfull, here is some relevant information about this request:"
          },
          "body": {
            "type": "string",
            "description": "A message about the operation or a token, in case of loggin an user.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "description": "Indicates whether an error occurred or not.",
            "example": true
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code.",
            "example": 500
          },
          "message": {
            "type": "string",
            "description": "A message describing what's the cause of the error. It varies a lot depending on the endpoint called and the different bodies provided.",
            "example": "Somewthing wrong ocurred, please try again."
          }
        }
      },
...
      "NotFoundResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "description": "Indicates whether an error occurred or not.",
            "example": true
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code.",
            "example": 404
          },
          "message": {
            "type": "string",
            "description": "The resource you were searching for wasn't found.",
            "example": "The resource X wasn't found or doesn't exists."
          }
        }
      },
...
    },
    "responses": {
...
      "InternalServerError": {
        "description": "There was an unexpected error during the request and couldn't perform the action.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}